+++ gridtools/__init__.py
@@ -10,4 +10,4 @@ Released under the terms of the BSD 3-Clause license.
 
 from .__info__ import __version__  # noqa: F401
 
-from .gridtools import Grid, LatLong, check_grid, check_latlong, distance  # noqa: F401
+from .gridtools import Grid, LatLong, check_grid, check_latlong, grid_distance  # noqa: F401