+++ CHANGELOG.md
@@ -11,12 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `LatLong` class to represent a latitude/longitude pair.
 - `check_grid()` to check if a string is a valid grid.
 - `check_latlong()` to check if a pair is a valid lat/long pair.
-- `distance()` to find the distance and bearing between two points.
+- `grid_distance()` to find the distance and bearing between two points.
 - A command-line interface.
 
 
 <!-- ## [1.0.0] - YYYY-MM-DD -->
 
 
-[Unreleased]: https://github.com/miaowware/gridtools/compare/1.0.0...HEAD
+<!-- [Unreleased]: https://github.com/miaowware/gridtools/compare/1.0.0...HEAD -->
 <!-- [1.0.0]: https://github.com/miaowware/gridtools/releases/tag/1.0.0 -->